refactor(MoveCountValidator): 메서드 클래스 변경
