+++ src/main/java/util/Validator.java
@@ -1,24 +1,5 @@
 package util;
 
-import java.util.regex.Pattern;
-
 public abstract class Validator {
-
-    private static final Pattern NUMBER_REGEX = Pattern.compile("^[0-9]*$");
-
     abstract void validate(String input) throws IllegalArgumentException;
-
-    void validateNumeric(String input) {
-        if (!NUMBER_REGEX.matcher(input).matches()) {
-            throw new IllegalArgumentException("자연수만 입력 가능합니다.");
-        }
-    }
-
-    void validateRange(String input) {
-        try {
-            Integer.parseInt(input);
-        } catch (NumberFormatException exception) {
-            throw new IllegalArgumentException("int 범위를 초과하는 입력은 불가능합니다.");
-        }
-    }
 }
